requirements.txt
numpy~=1.24.3
scipy
scikit-learn~=1.3.2
matplotlib
pandas~=2.0.2
fastapi~=0.108.0
uvicorn~=0.25.0
gunicorn==19.9.0
uvloop
httptools
pydantic~=1.10.4
scikit-surprise